Move all confd backends to the backends package
